re-factor requests
